Decoratorで書き換え
