+++ Gemfile
@@ -24,6 +24,7 @@ gem 'jbuilder', '~> 2.7'
 gem 'devise'
 gem 'hamlit'
 gem 'aws-sdk-s3', require: false
+gem 'active_decorator'
 
 # Use Active Storage variant
 gem 'image_processing', '~> 1.2'
